reset back the component in one file beside change technicalTask to be descriptive #31
